requirements.txt
# Django
Django==3.1.5
django-compressor==2.4.1
django-heroku==0.3.1
django-libsass==0.9
django-summernote==0.8.11.6
django-cloudinary-storage==0.3.0

# Django deployment
python-decouple==3.4
gunicorn==20.0.4
whitenoise==5.1.0

# Clean HTML
lxml==4.6.3
Markdown==3.3.4
markdownify==0.9.4
mdx-linkify==2.1
